Add files via upload
Title: "HTML-CSS YouTube Interface Clone: A Beginner's Journey"

Description:
This project is a beginner-level exploration into the world of web development, focusing on creating a basic interface clone of YouTube using HTML and CSS. Inspired by tutorials and learning materials found on YouTube, this endeavor serves as both a practical application of web development fundamentals and a tribute to the platform that made this learning journey possible. Throughout this project, I have aimed to capture the essence of YouTube's user interface, incorporating elements such as the navigation bar, video grid layout, and basic responsiveness to provide a simplified yet recognizable version of the popular video-sharing website. This project is intended for educational purposes, demonstrating the power of HTML and CSS in web interface design while acknowledging the vast learning resources available on YouTube for aspiring developers.